backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Marshmallow==1.1.0
marshmallow-sqlalchemy==0.29.0
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
PyJWT==2.8.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0